Refactored trendify API
