+++ src/trendify/__init__.py
@@ -2,5 +2,5 @@
 Provides top-level imports
 """
 
-from trendify.API import *
+from trendify.api import *
 from trendify.server import *
