+++ src/trendify/api/__init__.py
@@ -1,0 +1,9 @@
+from trendify.api.generator import *
+from trendify.api.plottable_data import *
+
+from trendify.api.api import *
+from trendify.api.data_product_collection import *
+from trendify.api.data_product import *
+from trendify.api.format2d import *
+from trendify.api.helpers import *
+from trendify.api.table import *
