+++ src/trendify/api/generator/__init__.py
@@ -1,0 +1,4 @@
+from trendify.api.generator.data_product_generator import *
+from trendify.api.generator.histogrammer import *
+from trendify.api.generator.table_builder import *
+from trendify.api.generator.xy_data_plotter import *
