+++ src/trendify/plotly_dashboard.py
@@ -16,20 +16,20 @@ from collections import defaultdict
 import os
 
 # Import from trendify
-from trendify.API import (
-    DataProductCollection,
-    Tag,
-    Tags,
-    Point2D,
-    Trace2D,
-    TableEntry,
-    HistogramEntry,
-    AxLine,
-    LineOrientation,
-    Format2D,
-    flatten,
-    atleast_1d,
-)
+# from trendify.api.API import (
+#     DataProductCollection,
+#     Tag,
+#     Tags,
+#     Point2D,
+#     Trace2D,
+#     TableEntry,
+#     HistogramEntry,
+#     AxLine,
+#     LineOrientation,
+#     Format2D,
+#     flatten,
+#     atleast_1d,
+# )
 
 # Constants
 COLORS = [
